Updated pydantic model generation in edittsv lifecycle
