+++ src/ipyautoui/automapschema.py
@@ -12,62 +12,13 @@ from ipyautoui.custom.markdown_widget import MarkdownWidget
 from ipyautoui.custom.filechooser import FileChooser
 from ipyautoui.custom.date_string import DatePickerString, NaiveDatetimePickerString
 from ipyautoui.autobox import AutoBox
-from tempfile import TemporaryDirectory
-import pathlib
-from datamodel_code_generator import DataModelType, InputFileType, generate
-import json
-import importlib.util
-import sys
 
 import logging
 
 logger = logging.getLogger(__name__)
 
-def pydantic_model_file_from_json_schema(json_schema, fpth):
-    return generate(
-            json.dumps(json_schema, ensure_ascii=False),
-            input_file_type=InputFileType.JsonSchema,
-            input_filename="example.json",
-            output=fpth,
-            output_model_type=DataModelType.PydanticV2BaseModel,
-            capitalise_enum_members=True,
-            field_include_all_keys=True
-        )
-
-def pydantic_model_from_json_schema(json_schema: dict) -> ty.Type[BaseModel]:
-    load = json_schema["title"].replace(" ", "") if "title" in json_schema else "Model"
-
-    with TemporaryDirectory() as temporary_directory_name:
-        temporary_directory = pathlib.Path(temporary_directory_name)
-        file_path = "model.py"
-        module_name = file_path.split(".")[0]
-        output = pathlib.Path(temporary_directory / file_path)
-        
-        pydantic_model_file_from_json_schema(json_schema, output)
-
-        #HACK refer to https://github.com/koxudaxi/datamodel-code-generator/issues/2534 for official fix, then remove the PATCH LOGIC once that is resolved
-        # --- NEW PATCH LOGIC ---
-        if json_schema.get("title") == "Project Building Area":
-            text = output.read_text()
-
-            # Replace Enum → IntEnum in TargetYear only
-            text = text.replace("class TargetYear(Enum):", "class TargetYear(IntEnum):")
-
-            # Ensure IntEnum is imported
-            if "from enum import IntEnum" not in text:
-                text = text.replace("from enum import Enum", "from enum import Enum, IntEnum")
-
-            output.write_text(text)
-        # --- END PATCH LOGIC ---
-
-        spec = importlib.util.spec_from_file_location(module_name, output)
-        module = importlib.util.module_from_spec(spec)
-        sys.modules[module_name] = module
-        spec.loader.exec_module(module)
-    return getattr(module, load)
-
 def _init_model_schema(
-    schema=None, by_alias=False, generate_pydantic_model_from_json_schema = True
+    schema=None, by_alias=False
 ) -> tuple[ty.Optional[ty.Type[BaseModel]], dict]:
     if schema is None:
         return None, {
@@ -76,17 +27,9 @@ def _init_model_schema(
             "items": {"properties": {}},
         }
     if isinstance(schema, dict):
-        if generate_pydantic_model_from_json_schema:
-            schema = replace_refs(schema, merge_props=True)
-            schema = {k: v for k, v in schema.items() if k != "$defs"}
-            model = pydantic_model_from_json_schema(schema)
-        else:
-            model = None
-            schema = replace_refs(schema, merge_props=True)
-            schema = {k: v for k, v in schema.items() if k != "$defs"}
-        # IDEA: Possible implementations -@jovyan at 8/24/2022, 12:05:02 PM
-        # jsonschema_to_pydantic
-        # https://koxudaxi.github.io/datamodel-code-generator/using_as_module/
+        model = None
+        schema = replace_refs(schema, merge_props=True)
+        schema = {k: v for k, v in schema.items() if k != "$defs"}
     else:
         model = schema  # the "model" passed is a pydantic model
         schema = model.model_json_schema(by_alias=by_alias).copy()
