+++ src/ipyautoui/custom/autogrid.py
@@ -514,7 +514,6 @@ class AutoGrid(DataGrid):
         by_alias: bool = False,
         by_title: bool = True,
         order: ty.Optional[tuple] = None,
-        generate_pydantic_model_from_json_schema: bool = False,
         **kwargs,
     ):
         self.__init__(
@@ -523,7 +522,6 @@ class AutoGrid(DataGrid):
             by_alias=by_alias,
             by_title=by_title,
             order=order,
-            generate_pydantic_model_from_json_schema=generate_pydantic_model_from_json_schema,
             **kwargs,
         )
 
@@ -606,7 +604,6 @@ class AutoGrid(DataGrid):
         by_alias: bool = False,
         by_title: bool = True,
         order: ty.Optional[tuple] = None,
-        generate_pydantic_model_from_json_schema: bool = False,
         **kwargs,
     ):
         # accept schema or pydantic schema
@@ -615,7 +612,7 @@ class AutoGrid(DataGrid):
         )
         self.by_title = by_title
         self.selection_mode = MAP_TRANSPOSED_SELECTION_MODE[self.transposed]
-        self.model, self.schema = asch._init_model_schema(schema, by_alias=by_alias, generate_pydantic_model_from_json_schema=generate_pydantic_model_from_json_schema)
+        self.model, self.schema = asch._init_model_schema(schema, by_alias=by_alias)
         # ^ generates gridschema
         self.gridschema.get_traits = self.datagrid_trait_names
         _data = self._init_data(data)
