+++ src/ipyautoui/demo_schemas/editable_datagrid.py
@@ -29,7 +29,6 @@ class EditableGrid(RootModel):
             # default_factory=lambda: DATAGRID_TEST_VALUE, # TODO: AutoUi isn't getting data when set using default_factory. make this work!
             format="DataFrame",
             show_ui_io=True,
-            generate_pydantic_model_from_json_schema=True,
             warn_on_delete=True,  # TODO: this isn't being passed
             global_decimal_places=2,
         )
